refactor(app): dedup backend commands and kill dead code
- drag_ops: shared resolve_drag_image/launch_drag/on_drop_hide_windows; drop unused State params + FileList alias + dead badge-count code

- file: FileMetadata::from_path shared by add_files and file_drop; add_files skips unreadable paths instead of failing batch

- file_ops: save_pasted_* delegate to save_bytes_to_drop_folder; remove_files emits once per batch

- window_ops: reveal_main_window/open_updater_window shared by hotkey/tray/shortcut/updater paths (adds missing unminimize)

- mime: shared ext_from_image_mime for download + data-URL paths

- config_ops: delete macOS keycode table, input-monitoring stubs; merge register_hotkey; inline config_exists

- misc: drop extract_text wrapper, get_hbitmap inline, dead position fetch, commented handler, triple-Arc DragState; mouse_monitor/win.rs folds into mod.rs
