+++ app/src-tauri/src/commands/drag_ops.rs
@@ -1,16 +1,87 @@
-use crate::config::AppConfig;
-use crate::file::FileMetadata;
-use std::sync::{Arc, Mutex};
-use tauri::{AppHandle, Manager, State};
+use tauri::{AppHandle, Manager};
 use tracing::{error, info, warn};
 
-type FileList = Arc<Mutex<Vec<FileMetadata>>>;
+/// Decode the frontend-provided drag image (`data:...;base64,...` or raw
+/// base64), falling back to a generated badge image for `fallback_count`.
+fn resolve_drag_image(drag_image: Option<String>, fallback_count: usize) -> drag::Image {
+    if let Some(base64_data) = drag_image {
+        // Remove data URL prefix if present (e.g., "data:image/png;base64,")
+        let base64_str = if let Some(comma_pos) = base64_data.find(',') {
+            &base64_data[comma_pos + 1..]
+        } else {
+            &base64_data
+        };
+
+        match base64::Engine::decode(&base64::engine::general_purpose::STANDARD, base64_str) {
+            Ok(bytes) => {
+                info!("Using frontend-provided drag image ({} bytes)", bytes.len());
+                return drag::Image::Raw(bytes);
+            }
+            Err(e) => {
+                warn!(
+                    "Failed to decode drag image, falling back to generated image: {}",
+                    e
+                );
+            }
+        }
+    }
+    generate_drag_image(fallback_count)
+}
+
+/// After a successful native drop the shelf UI gets out of the way:
+/// the popup (if open) closes and the main window hides.
+fn on_drop_hide_windows(app: &AppHandle, result: drag::DragResult) {
+    if matches!(result, drag::DragResult::Cancel) {
+        return;
+    }
+
+    // check if the popup window is open
+    if app.get_webview_window("popup").is_some() {
+        if let Err(e) = super::window_ops::close_popup_window(app.clone()) {
+            error!("Failed to close popup window after drag: {}", e);
+        }
+    }
+    if let Some(main_window) = app.get_webview_window("main") {
+        if let Err(e) = main_window.hide() {
+            error!("Failed to hide main window after drag: {}", e);
+        }
+    }
+}
+
+fn launch_drag(app: &AppHandle, item: drag::DragItem, image: drag::Image) -> Result<(), String> {
+    let window = app
+        .get_webview_window("main")
+        .ok_or("Main window not found")?;
+
+    let app_clone = app.clone();
+    let on_drop_callback = move |result: drag::DragResult, _: drag::CursorPosition| {
+        on_drop_hide_windows(&app_clone, result);
+    };
+
+    match drag::start_drag(
+        &window,
+        item,
+        image,
+        on_drop_callback,
+        drag::Options {
+            skip_animatation_on_cancel_or_failure: true,
+            mode: drag::DragMode::CopyOrMove,
+        },
+    ) {
+        Ok(_) => {
+            info!("Native drag started successfully");
+            Ok(())
+        }
+        Err(e) => {
+            error!("Failed to start native drag: {:?}", e);
+            Err(format!("Failed to start drag operation: {:?}", e))
+        }
+    }
+}
 
 #[tauri::command]
 pub fn start_multi_drag(
     app: AppHandle,
-    _file_list: State<'_, FileList>,
-    _config: State<'_, Arc<Mutex<AppConfig>>>,
     file_paths: Vec<String>,
     drag_image: Option<String>,
 ) -> Result<(), String> {
@@ -37,88 +108,14 @@ pub fn start_multi_drag(
         return Err("No valid files to drag".to_string());
     }
 
-    // Use the drag image from the frontend if provided, otherwise generate one
-    let image = if let Some(base64_data) = drag_image {
-        // Remove data URL prefix if present (e.g., "data:image/png;base64,")
-        let base64_str = if let Some(comma_pos) = base64_data.find(',') {
-            &base64_data[comma_pos + 1..]
-        } else {
-            &base64_data
-        };
-
-        match base64::Engine::decode(&base64::engine::general_purpose::STANDARD, base64_str) {
-            Ok(bytes) => {
-                info!("Using frontend-provided drag image ({} bytes)", bytes.len());
-                drag::Image::Raw(bytes)
-            }
-            Err(e) => {
-                warn!(
-                    "Failed to decode drag image, falling back to generated image: {}",
-                    e
-                );
-                generate_drag_image(valid_paths.len())
-            }
-        }
-    } else {
-        generate_drag_image(valid_paths.len())
-    };
-
+    let image = resolve_drag_image(drag_image, valid_paths.len());
     let item = drag::DragItem::Files(valid_paths.clone());
     info!(
         "Prepared drag item with {} valid file(s)",
         valid_paths.len()
     );
 
-    let window = app
-        .get_webview_window("main")
-        .ok_or("Main window not found")?;
-
-
-
-    let app_clone = app.clone();
-
-    let on_drop_callback = move |result: drag::DragResult, _: drag::CursorPosition| {
-        if matches!(result, drag::DragResult::Cancel) {
-            return;
-        }
-
-        // check if the popup window is open
-        if app_clone.get_webview_window("popup").is_some() {
-            if let Err(e) = super::window_ops::close_popup_window(app_clone.clone()) {
-                error!("Failed to close popup window after drag: {}", e);
-            }
-        }
-        if let Some(main_window) = app_clone.get_webview_window("main") {
-            if let Err(e) = main_window.hide() {
-                error!("Failed to hide main window after drag: {}", e);
-            }
-        }
-    };
-
-    let mode = drag::DragMode::CopyOrMove;
-
-    match drag::start_drag(
-        &window,
-        item,
-        image,
-        on_drop_callback,
-        drag::Options {
-            skip_animatation_on_cancel_or_failure: true,
-            mode,
-        },
-    ) {
-        Ok(_) => {
-            info!("Native drag started successfully");
-            Ok(())
-        }
-        Err(e) => {
-            error!("Failed to start native drag: {:?}", e);
-            Err(format!(
-                "Failed to start multi-file drag operation: {:?}",
-                e
-            ))
-        }
-    }
+    launch_drag(&app, item, image)
 }
 
 /// Generate a simple drag image with file count badge using the `image` crate.
@@ -178,13 +175,14 @@ fn generate_drag_image(file_count: usize) -> drag::Image {
         }
     }
 
-    // If multiple files, draw a badge circle with count
+    // If multiple files, draw a badge circle. The badge alone signals
+    // "multiple" — the `image` crate has no text rasterization, so no count
+    // is drawn inside it.
     if file_count > 1 {
         let badge_radius = 18i32;
         let badge_cx = (size - margin) as i32;
         let badge_cy = (size - margin) as i32;
         let badge_color = Rgba([59, 130, 246, 255]); // Blue
-        let _badge_text_color = Rgba([255, 255, 255, 255]);
 
         // Draw badge circle
         for y in 0..size as i32 {
@@ -196,18 +194,6 @@ fn generate_drag_image(file_count: usize) -> drag::Image {
                 }
             }
         }
-
-        // Draw count number (simple pixel art for single/double digit)
-        let count_str = if file_count > 99 {
-            "99+".to_string()
-        } else {
-            file_count.to_string()
-        };
-        // For simplicity, just draw a small dot pattern - the badge itself is informative
-        let _ = count_str; // Count shown by badge presence; actual text rendering is complex with `image` crate
-
-        // Draw a simple "+" or number shape in the badge center
-        // For now, just the badge alone indicates multiple files
     }
 
     // Encode to PNG
@@ -232,24 +218,7 @@ pub fn start_text_drag(
 ) -> Result<(), String> {
     info!("Starting text drag");
 
-    // Use the drag image from the frontend if provided, otherwise generate one
-    let image = if let Some(base64_data) = drag_image {
-        let base64_str = if let Some(comma_pos) = base64_data.find(',') {
-            &base64_data[comma_pos + 1..]
-        } else {
-            &base64_data
-        };
-
-        match base64::Engine::decode(&base64::engine::general_purpose::STANDARD, base64_str) {
-            Ok(bytes) => drag::Image::Raw(bytes),
-            Err(e) => {
-                warn!("Failed to decode drag image, falling back: {}", e);
-                generate_drag_image(1)
-            }
-        }
-    } else {
-        generate_drag_image(1)
-    };
+    let image = resolve_drag_image(drag_image, 1);
 
     let text_clone = text.clone();
     let provider: drag::DataProvider = Box::new(move |format: &str| -> Option<Vec<u8>> {
@@ -265,48 +234,5 @@ pub fn start_text_drag(
         types: vec!["text/plain".to_string()],
     };
 
-    let window = app
-        .get_webview_window("main")
-        .ok_or("Main window not found")?;
-
-
-
-    let app_clone = app.clone();
-    let on_drop_callback = move |result: drag::DragResult, _: drag::CursorPosition| {
-        if matches!(result, drag::DragResult::Cancel) {
-            return;
-        }
-        if app_clone.get_webview_window("popup").is_some() {
-            if let Err(e) = super::window_ops::close_popup_window(app_clone.clone()) {
-                error!("Failed to close popup window after drag: {}", e);
-            }
-        }
-        if let Some(main_window) = app_clone.get_webview_window("main") {
-            if let Err(e) = main_window.hide() {
-                error!("Failed to hide main window after drag: {}", e);
-            }
-        }
-    };
-
-    let mode = drag::DragMode::CopyOrMove;
-
-    match drag::start_drag(
-        &window,
-        item,
-        image,
-        on_drop_callback,
-        drag::Options {
-            skip_animatation_on_cancel_or_failure: true,
-            mode,
-        },
-    ) {
-        Ok(_) => {
-            info!("Native text drag started successfully");
-            Ok(())
-        }
-        Err(e) => {
-            error!("Failed to start native text drag: {:?}", e);
-            Err(format!("Failed to start text drag operation: {:?}", e))
-        }
-    }
+    launch_drag(&app, item, image)
 }
